:recycle: cdnURL rebuild
