+++ src/main/java/com/tale/controller/admin/IndexController.java
@@ -1,21 +1,31 @@
 package com.tale.controller.admin;
 
 import com.blade.ioc.annotation.Inject;
+import com.blade.kit.DateKit;
 import com.blade.mvc.annotation.GetRoute;
+import com.blade.mvc.annotation.JSON;
 import com.blade.mvc.annotation.Path;
+import com.blade.mvc.annotation.PostRoute;
 import com.blade.mvc.http.Request;
+import com.blade.mvc.multipart.FileItem;
+import com.blade.mvc.ui.RestResponse;
+import com.tale.annotation.SysLog;
+import com.tale.bootstrap.TaleConst;
 import com.tale.controller.BaseController;
 import com.tale.model.dto.Statistics;
 import com.tale.model.dto.Types;
-import com.tale.model.entity.Comments;
-import com.tale.model.entity.Contents;
-import com.tale.model.entity.Logs;
+import com.tale.model.entity.*;
 import com.tale.service.OptionsService;
 import com.tale.service.SiteService;
+import com.tale.utils.TaleUtils;
 import io.github.biezhi.anima.enums.OrderBy;
 import io.github.biezhi.anima.page.Page;
 import lombok.extern.slf4j.Slf4j;
 
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
@@ -71,5 +81,63 @@ public class IndexController extends BaseController {
     public String profile() {
         return "admin/profile";
     }
+
+    /**
+     * 上传文件接口
+     */
+    @SysLog("上传附件")
+    @PostRoute("api/attach/upload")
+    @JSON
+    public RestResponse<?> upload(Request request) {
+
+        log.info("UPLOAD DIR = {}", TaleUtils.UP_DIR);
+
+        Users        users      = this.user();
+        Integer      uid        = users.getUid();
+        List<Attach> errorFiles = new ArrayList<>();
+        List<Attach> urls       = new ArrayList<>();
+
+        Map<String, FileItem> fileItems = request.fileItems();
+        if (null == fileItems || fileItems.size() == 0) {
+            return RestResponse.fail("请选择文件上传");
+        }
+
+        fileItems.forEach((fileName, fileItem) -> {
+            String fname = fileItem.getFileName();
+            if ((fileItem.getLength() / 1024) <= TaleConst.MAX_FILE_SIZE) {
+                String fkey = TaleUtils.getFileKey(fname);
+
+                String ftype    = fileItem.getContentType().contains("image") ? Types.IMAGE : Types.FILE;
+                String filePath = TaleUtils.UP_DIR + fkey;
+
+                try {
+                    Files.write(Paths.get(filePath), fileItem.getData());
+                } catch (IOException e) {
+                    log.error("", e);
+                }
+
+                Attach attach = new Attach();
+                attach.setFname(fname);
+                attach.setAuthorId(uid);
+                attach.setFkey(fkey);
+                attach.setFtype(ftype);
+                attach.setCreated(DateKit.nowUnix());
+                attach.save();
+
+                urls.add(attach);
+                siteService.cleanCache(Types.C_STATISTICS);
+            } else {
+                Attach attach = new Attach();
+                attach.setFname(fname);
+                errorFiles.add(attach);
+            }
+        });
+
+        if (errorFiles.size() > 0) {
+            return RestResponse.fail().payload(errorFiles);
+        }
+        return RestResponse.ok(urls);
+    }
+
 
 }
