+++ src/main/java/com/tale/controller/admin/TemplateController.java
@@ -5,8 +5,6 @@ import com.blade.mvc.Const;
 import com.blade.mvc.annotation.Param;
 import com.blade.mvc.annotation.Path;
 import com.blade.mvc.annotation.PostRoute;
-import com.blade.mvc.annotation.Route;
-import com.blade.mvc.http.HttpMethod;
 import com.blade.mvc.ui.RestResponse;
 import com.tale.annotation.SysLog;
 import com.tale.controller.BaseController;
