+++ src/main/java/com/tale/hooks/BaseWebHook.java
@@ -8,8 +8,6 @@ import com.blade.mvc.http.Request;
 import com.blade.mvc.http.Response;
 import com.tale.annotation.SysLog;
 import com.tale.bootstrap.TaleConst;
-import com.tale.extension.Commons;
-import com.tale.model.dto.Types;
 import com.tale.model.entity.Logs;
 import com.tale.model.entity.Users;
 import com.tale.utils.TaleUtils;
@@ -68,10 +66,6 @@ public class BaseWebHook implements WebHook {
                 logs.save();
             }
         }
-
-        signature.request().attribute(Types.ATTACH_URL, Commons.site_option(Types.ATTACH_URL, Commons.site_url()));
-        signature.request().attribute("max_file_size", TaleConst.MAX_FILE_SIZE / 1024);
-
         return true;
     }
