+++ src/main/resources/templates/admin/article/_footer.html
@@ -1,0 +1,14 @@
+#include('../footer.html')
+<script src="${cdnURL()}/plugins/tagsinput/jquery.tagsinput.min.js"></script>
+<script src="${cdnURL()}/plugins/datetimepicker/bootstrap-datetimepicker.min.js"></script>
+<script src="${cdnURL()}/plugins/datetimepicker/bootstrap-datetimepicker.zh-CN.js"></script>
+<script src="${cdnURL()}/plugins/jquery-multi-select/jquery.quicksearch.js"></script>
+<script src="${cdnURL()}/plugins/mditor/js/mditor.min.js"></script>
+<script src="${cdnURL()}/plugins/wysihtml5/0.3.0/wysihtml5.min.js"></script>
+<script src="${cdnURL()}/plugins/summernote/0.8.2/summernote.min.js"></script>
+<script src="${cdnURL()}/plugins/summernote/0.8.2/lang/summernote-zh-CN.min.js"></script>
+<script src="${cdnURL()}/plugins/multi-select/0.9.12/js/jquery.multi-select.min.js"></script>
+<script src="${cdnURL()}/plugins/select2/3.4.8/select2.min.js"></script>
+<script src="${cdnURL()}/plugins/jquery-toggles/2.0.4/toggles.min.js"></script>
+<script src="${cdnURL()}/plugins/dropzone/4.3.0/min/dropzone.min.js"></script>
+<script src="${cdnURL()}/plugins/to-markdown/3.1.0/to-markdown.min.js"></script>
