+++ src/main/resources/templates/admin/article/_header.html
@@ -1,0 +1,76 @@
+<link href="${cdnURL()}/plugins/tagsinput/jquery.tagsinput.css" rel="stylesheet">
+<link href="${cdnURL()}/plugins/select2/dist/css/select2-bootstrap.css" rel="stylesheet">
+<link href="${cdnURL()}/plugins/toggles/toggles.css" rel="stylesheet">
+<link href="${cdnURL()}/plugins/datetimepicker/bootstrap-datetimepicker.min.css" rel="stylesheet">
+<link href="${cdnURL()}/plugins/multi-select/0.9.12/css/multi-select.min.css" rel="stylesheet"/>
+<link href="${cdnURL()}/plugins/select2/3.4.8/select2.min.css" rel="stylesheet"/>
+<link href="${cdnURL()}/plugins/mditor/css/mditor.min.css" rel="stylesheet"/>
+<link href="${cdnURL()}/plugins/summernote/0.8.2/summernote.css" rel="stylesheet">
+<link href="${cdnURL()}/plugins/dropzone/4.3.0/min/dropzone.min.css" rel="stylesheet">
+
+<style rel="stylesheet">
+    #tags_tagsinput {
+        background-color: #fafafa;
+        border: 1px solid #eeeeee;
+    }
+
+    #tags_addTag input {
+        width: 100%;
+    }
+
+    #tags_addTag {
+        margin-top: -5px;
+    }
+
+    .mditor .editor{
+        font-size: 14px;
+        font-family: "Helvetica Neue",Helvetica,Arial,sans-serif;
+    }
+    .mditor .backdrop, .mditor .textarea, .mditor .viewer{
+        font-size: 14px;
+    }
+    .markdown-body{
+        font-size: 14px;
+    }
+    .note-toolbar {
+        text-align: center;
+    }
+
+    .note-editor.note-frame {
+        border: none;
+    }
+
+    .note-editor .note-toolbar {
+        background-color: #f5f5f5;
+        padding-bottom: 10px;
+    }
+
+    .note-toolbar .note-btn-group {
+        margin: 0;
+    }
+
+    .note-toolbar .note-btn {
+        border: none;
+    }
+
+    #articleForm #dropzone {
+        min-height: 200px;
+        background-color: #dbdde0;
+        line-height:200px;
+        margin-bottom: 10px;
+    }
+    #articleForm .dropzone {
+        border: 1px dashed #8662c6;
+        border-radius: 5px;
+        background: white;
+    }
+    #articleForm .dropzone .dz-message {
+        font-weight: 400;
+    }
+    #articleForm .dropzone .dz-message .note {
+        font-size: 0.8em;
+        font-weight: 200;
+        display: block;
+        margin-top: 1.4rem;
+    }
+</style>
