+++ src/main/resources/templates/admin/article/edit.html
@@ -1,82 +1,12 @@
 #include('../header.html',{active: 'publish', title:'编辑文章'})
-<link href="/static/admin/plugins/tagsinput/jquery.tagsinput.css" rel="stylesheet">
-<link href="/static/admin/plugins/select2/dist/css/select2-bootstrap.css" rel="stylesheet">
-<link href="/static/admin/plugins/toggles/toggles.css" rel="stylesheet">
-<link href="/static/admin/plugins/datetimepicker/bootstrap-datetimepicker.min.css" rel="stylesheet">
-#call cdn_url('article_edit_head')
-<style rel="stylesheet">
-    #tags_tagsinput {
-        background-color: #fafafa;
-        border: 1px solid #eeeeee;
-    }
-
-    #tags_addTag input {
-        width: 100%;
-    }
-
-    #tags_addTag {
-        margin-top: -5px;
-    }
-
-    .mditor .editor{
-        font-size: 14px;
-        font-family: "Helvetica Neue",Helvetica,Arial,sans-serif;
-    }
-    .mditor .backdrop, .mditor .textarea, .mditor .viewer{
-        font-size: 14px;
-    }
-    .markdown-body{
-        font-size: 14px;
-    }
-    .note-toolbar {
-        text-align: center;
-    }
-
-    .note-editor.note-frame {
-        border: none;
-    }
-
-    .note-editor .note-toolbar {
-        background-color: #f5f5f5;
-        padding-bottom: 10px;
-    }
-
-    .note-toolbar .note-btn-group {
-        margin: 0;
-    }
-
-    .note-toolbar .note-btn {
-        border: none;
-    }
-
-    #articleForm #dropzone {
-        min-height: 200px;
-        background-color: #dbdde0;
-        line-height:200px;
-        margin-bottom: 10px;
-    }
-    #articleForm .dropzone {
-        border: 1px dashed #8662c6;
-        border-radius: 5px;
-        background: white;
-    }
-    #articleForm .dropzone .dz-message {
-        font-weight: 400;
-    }
-    #articleForm .dropzone .dz-message .note {
-        font-size: 0.8em;
-        font-weight: 200;
-        display: block;
-        margin-top: 1.4rem;
-    }
-</style>
+#include('./_header.html')
 <div class="row">
     <div class="col-sm-12">
         <h4 class="page-title">编辑文章</h4>
     </div>
     <div id="app" class="col-md-12">
 
-        <input type="hidden" id="attach_url" value="${attach_url}" />
+        <input type="hidden" id="attach_url" value="${attachURL()}" />
 
         <form id="articleForm">
             <input type="hidden" v-model="article.cid"/>
@@ -185,12 +115,7 @@
         </form>
     </div>
 </div>
-#include('../footer.html')
-<script src="/static/admin/plugins/tagsinput/jquery.tagsinput.min.js"></script>
-<script src="/static/admin/plugins/datetimepicker/bootstrap-datetimepicker.min.js"></script>
-<script src="/static/admin/plugins/datetimepicker/bootstrap-datetimepicker.zh-CN.js"></script>
-<script src="/static/admin/plugins/jquery-multi-select/jquery.quicksearch.js"></script>
-#call cdn_url('article_edit_foot')
-<script src="/static/admin/js/edit_article.js?v=${version}" type="text/javascript"></script>
+#include('./_footer.html')
+<script src="${cdnURL()}/js/edit_article.js?v=${version}" type="text/javascript"></script>
 </body>
 </html>
