+++ src/main/resources/templates/admin/footer.html
@@ -3,9 +3,11 @@
 </footer>
 </div>
 </div>
-#call cdn_url('footer')
-<script src="/static/admin/plugins/axios/axios.min.js"></script>
-<script src="/static/admin/plugins/vue/vue.js"></script>
-<script src="/static/admin/plugins/moment/moment.min.js"></script>
-<script src="/static/admin/js/jquery.app.js?v=${version}"></script>
-<script src="/static/admin/js/base.js?v=${version}"></script>
+<script src="${cdnURL()}/plugins/jquery/3.2.1/jquery.min.js"></script>
+<script src="${cdnURL()}/plugins/bootstrap/3.3.7/js/bootstrap.min.js"></script>
+<script src="${cdnURL()}/plugins/limonte-sweetalert2/6.4.1/sweetalert2.min.js"></script>
+<script src="${cdnURL()}/plugins/axios/axios.min.js"></script>
+<script src="${cdnURL()}/plugins/vue/vue.js"></script>
+<script src="${cdnURL()}/plugins/moment/moment.min.js"></script>
+<script src="${cdnURL()}/js/jquery.app.js?v=${version}"></script>
+<script src="${cdnURL()}/js/base.js?v=${version}"></script>
