+++ src/main/resources/templates/admin/login.html
@@ -6,11 +6,14 @@
     <meta content="${_csrf_token}" name="csrf_token"/>
     <link rel="shortcut icon" href="/static/admin/images/favicon.png"/>
     <title>Tale - 用户登录</title>
-    #call cdn_url('login_head')
-    <link href="/static/admin/css/style.min.css?v=${version}" rel="stylesheet" type="text/css">
+    <link href="${cdnURL()}/plugins/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet">
+    <link href="${cdnURL()}/plugins/bootstrap/3.3.7/css/bootstrap.min.css" rel="stylesheet" type="text/css"/>
+    <link href="${cdnURL()}/plugins/limonte-sweetalert2/6.4.1/sweetalert2.min.css" rel="stylesheet">
+    <link href="${cdnURL()}/css/style.min.css?v=${version}" rel="stylesheet" type="text/css">
+
     <style type="text/css">
         body,html {
-            background: url("/static/admin/images/bg/${random(5, '.png')}") no-repeat;
+            background: url("${cdnURL()}/images/bg/${random(5, '.png')}") no-repeat;
             background-size: cover;
         }
         .panel-shadow {
@@ -46,7 +49,8 @@
     <!-- HTML5 Shim and Respond.js IE8 support of HTML5 elements and media queries -->
     <!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
     <!--[if lt IE 9]>
-    #call cdn_url('html5_respond')
+    <script src="${cdnURL()}/plugins/html5shiv/3.7.0/html5shiv.js"></script>
+    <script src="${cdnURL()}/plugins/respond.js/1.3.0/respond.min.js"></script>
     <![endif]-->
 
 </head>
@@ -97,8 +101,11 @@
     </div>
 </div>
 <!-- Main  -->
-#call cdn_url('login_foot')
-<script src="/static/admin/js/base.js"></script>
+<script src="${cdnURL()}/plugins/vue/vue.js"></script>
+<script src="${cdnURL()}/plugins/jquery/3.2.1/jquery.min.js"></script>
+<script src="${cdnURL()}/plugins/limonte-sweetalert2/6.4.1/sweetalert2.min.js"></script>
+<script src="${cdnURL()}/plugins/axios/axios.min.js"></script>
+<script src="${cdnURL()}/js/base.js"></script>
 <script type="text/javascript">
     var tale = new $.tale();
     function checkForm() {
