+++ src/main/resources/templates/admin/page_edit.html
@@ -41,7 +41,7 @@
     </div>
     <div class="col-md-12">
 
-        <input type="hidden" id="attach_url" value="${attach_url}" />
+        <input type="hidden" id="attach_url" value="${attachURL()}" />
 
         <form id="articleForm" role="form" novalidate="novalidate">
             <input type="hidden" name="cid" value="${contents.cid ?! ''}" id="cid"/>
