+++ src/main/resources/templates/admin/profile.html
@@ -81,7 +81,8 @@
 
 </div>
 #include('./footer.html')
-#call cdn_url('profile_foot')
+<script src="${cdnURL()}/plugins/jquery-validate/1.15.1/jquery.validate.min.js"></script>
+<script src="${cdnURL()}/plugins/jquery-validate/1.15.1/localization/messages_zh.min.js"></script>
 <script type="text/javascript">
     var tale = new $.tale();
