+++ src/main/resources/templates/install.html
@@ -7,7 +7,8 @@
     <link rel="shortcut icon" href="/static/admin/images/favicon.png"/>
     <title>Tale - 博客安装</title>
     <link href="/static/admin/plugins/jquery-steps/1.1.0/jquery.steps.css" rel="stylesheet" type="text/css">
-    #call cdn_url('install_head')
+    <link href="/static/admin/plugins/bootstrap/3.3.7/css/bootstrap.min.css" rel="stylesheet" type="text/css"/>
+    <link href="/static/admin/plugins/limonte-sweetalert2/6.4.1/sweetalert2.min.css" rel="stylesheet">
     <link href="/static/admin/css/style.min.css" rel="stylesheet" type="text/css">
     <style>
         body,html {
@@ -138,7 +139,14 @@
     文件。
     #end
 </div>
-#call cdn_url('install_foot')
+<script src="/static/admin/plugins/jquery/3.2.1/jquery.min.js"></script>
+<script src="/static/admin/plugins/bootstrap/3.3.7/js/bootstrap.min.js"></script>
+<script src="/static/admin/plugins/vue/vue.min.js"></script>
+<script src="/static/admin/plugins/axios/axios.min.js"></script>
+<script src="/static/admin/plugins/limonte-sweetalert2/6.4.1/sweetalert2.min.js"></script>
+<script src="/static/admin/plugins/jquery-validate/1.15.1/jquery.validate.min.js"></script>
+<script src="/static/admin/plugins/jquery-validate/1.15.1/localization/messages_zh.min.js"></script>
+<script src="/static/admin/plugins/jquery-steps/1.1.0/jquery.steps.min.js"></script>
 <script src="/static/admin/js/base.js?v=${version}"></script>
 <script src="/static/admin/js/install.js?v=${version}"></script>
 </body>
