+++ src/main/resources/templates/themes/default/partial/header.html
@@ -15,7 +15,7 @@
     <title>${head_title()}</title>
     <link href="${theme_url('/static/css/xcode.min.css')}" rel="stylesheet">
     <link href="${theme_url('/static/css/style.min.css')}" rel="stylesheet">
-    #call cdn_url("jquery")
+    <script src="/static/admin/plugins/jquery/3.2.1/jquery.min.js"></script>
 
     <!--[if lt IE 9]>
     <script src="//cdn.bootcss.com/html5shiv/r29/html5.min.js"></script>
